案件詳細画面(作業中) 作業完了ボタンの実装

+++ app/src/main/java/jp/co/recruit/erikura/data/network/IErikuraApiService.kt
@@ -253,7 +253,7 @@ data class StopJobRequest(
     var latitude: Double,
     var longitude: Double,
     var steps: Int,
-    var distance: Int,
+    var distance: Double,
     var floorAsc: Int,
     var floorDesc: Int
 )
